Add Degree entity and update User and Course entities
This commit adds a new entity called Degree to the Course module. It also updates the User and Course entities to include a OneToMany relationship with Degree. The Degree entity has columns for title, degree, and foreign keys to Course and User entities. This update allows for tracking of degrees earned by students in a particular course.