Refactor user creation methods to handle both doctors and teacher assistants
Previously, the `createDR` method in `AuthService` was responsible for creating doctor users, and a separate `createTA` method was intended for creating teacher assistant users. This commit refactors the `createDR` method to handle both types of users by renaming it to `createDRorTA` and modifying its signature to accept either a `CreateDoctorDto` or a `CreateTeacherAssistUserDto`.

Additionally, the method now uses the `userRepository.create` method to initialize the user object, which simplifies the code and improves readability. The `createHOD` method remains unchanged and continues to handle the creation of Head of Department users.

This change is propagated to the `DoctorService` and `TeacherassistService` where the `create` methods now call the updated `createDRorTA` method in `AuthService`. This ensures that both doctor and teacher assistant users are created using the same logic, reducing code duplication and improving maintainability.